refactor: remove custom event with device context state
