+++ app/_features/room/components/event-container.tsx
@@ -1,7 +1,6 @@
 'use client';
 import { useEffect } from 'react';
 import { usePeerContext } from '@/_features/room/contexts/peer-context';
-import { hasTouchScreen } from '@/_shared/utils/has-touch-screen';
 
 declare global {
   interface Window {
@@ -20,14 +19,6 @@ export default function EventContainer({
   useEffect(() => {
     if (!peer) return;
 
-    const onTurnOnCamera = () => {
-      if (peer) peer.turnOnCamera();
-    };
-
-    const onTurnOffCamera = () => {
-      if (peer) peer.turnOffCamera();
-    };
-
     const onTurnOnMic = () => {
       if (peer) peer.turnOnMic();
     };
@@ -36,35 +27,12 @@ export default function EventContainer({
       if (peer) peer.turnOffMic();
     };
 
-    document.addEventListener('trigger:turnon-camera', onTurnOnCamera);
-    document.addEventListener('trigger:turnoff-camera', onTurnOffCamera);
     document.addEventListener('trigger:turnon-mic', onTurnOnMic);
     document.addEventListener('trigger:turnoff-mic', onTurnOffMic);
 
     return () => {
-      document.removeEventListener('trigger:turnon-camera', onTurnOnCamera);
-      document.removeEventListener('trigger:turnoff-camera', onTurnOffCamera);
       document.removeEventListener('trigger:turnon-mic', onTurnOnMic);
       document.removeEventListener('trigger:turnoff-mic', onTurnOffMic);
-    };
-  }, [peer]);
-
-  useEffect(() => {
-    if (!peer) return;
-
-    const isTouchScreen = hasTouchScreen();
-
-    const onWindowBlur = () => {
-      if (isTouchScreen && peer) {
-        document.dispatchEvent(new CustomEvent('trigger:turnoff-camera'));
-        document.dispatchEvent(new CustomEvent('trigger:turnoff-mic'));
-      }
-    };
-
-    window.addEventListener('blur', onWindowBlur);
-
-    return () => {
-      window.removeEventListener('blur', onWindowBlur);
     };
   }, [peer]);
